requirements.txt
albucore==0.0.21
albumentations==1.4.23
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
Django==5.1.4
django-allauth==65.1.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
efficientnet_pytorch==0.7.1
eval_type_backport==0.2.0
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
flake8==7.1.1
fonttools==4.53.1
fsspec==2024.10.0
gensim==4.3.3
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pep8-naming==0.13.3
pillow==10.4.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
prettytable==3.11.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
psutil==6.0.0
pure_eval==0.2.3
pyasn1-modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306; sys_platform == 'win32'
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
selenium==4.29.0
simsimd==6.2.1
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
stringzilla==3.11.1
sympy==1.13.1
threadpoolctl==3.5.0
timm==0.9.7
tokenizers==0.21.0
torch==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wrapt==1.17.0
wsproto==1.2.0
django-phonenumber-field[phonenumbers]
gunicorn==20.1.0